+++ ggFrames/ggSavedVarsDefaults.lua
@@ -1,22 +1,27 @@
 GGF.SavedVarsDefaults = {
+  -- Player
   ["PlayerContainer_OffsetX"] = 3,
   ["PlayerContainer_OffsetY"] = 3,
+  
   ["Player_Width"] = 300,
   ["Player_Font"] = "Bebas Neue",
   ["Player_FontSize"] = 18,
   ["Player_FontColor"] = {1,1,1,1},
+  
   ["Player_Health_Height"] = 36,
   ["Player_Health_BarColor"] = {0.831372549,0.149019608,0.149019608,1},
   ["Player_Health_Texture"] = "Round",
   ["Player_Health_Font"] = "Bebas Neue",
   ["Player_Health_FontSize"] = 16, 
   ["Player_Health_FontColor"] = {1,1,1,1},
+  
   ["Player_Magicka_Height"] = 20,
   ["Player_Magicka_BarColor"] = {.08627451,.294117647,.776470588,1},
   ["Player_Magicka_Texture"] = "Round",
   ["Player_Magicka_Font"] = "Bebas Neue",
   ["Player_Magicka_FontSize"] = 16, 
   ["Player_Magicka_FontColor"] = {1,1,1,1},
+  
   ["Player_Stamina_Height"] = 20,
   ["Player_Stamina_BarColor"] = {.796078431,.643137255,.043137255,1},
   ["Player_Stamina_Texture"] = "Round",
@@ -28,7 +33,6 @@ GGF.SavedVarsDefaults = {
   ["Player_Experience_Height"] = 8,
   ["Player_Experience_BarColor"] = {.7,.7,.7,1},
   ["Player_Experience_Texture"] = "Round",
-  
   ["Player_Experience_Text_Display"] = false,
   ["Player_Experience_Font"] = "Bebas Neue",
   ["Player_Experience_FontSize"] = 14, 
@@ -38,12 +42,15 @@ GGF.SavedVarsDefaults = {
   ["Player_Mount_BarColor"] = {.933333333,.933333333,0,1},
   ["Player_Mount_Texture"] = "Round",
 
+  -- GROUP
   ["GroupContainer_OffsetX"] = 3,
   ["GroupContainer_OffsetY"] = 140,
+ 
   ["Group_Width"] = 300,
   ["Group_Font"] = "Bebas Neue",
   ["Group_FontSize"] = 16,
   ["Group_FontColor"] = {1,1,1,1},
+ 
   ["Group_Health_Height"] = 36,
   ["Group_Health_BarColor"] = {0.831372549,0.149019608,0.149019608,1},
   ["Group_Health_Texture"] = "Round",
@@ -51,13 +58,35 @@ GGF.SavedVarsDefaults = {
   ["Group_Health_FontSize"] = 14, 
   ["Group_Health_FontColor"] = {1,1,1,1},
 
+  -- RAID
   ["LargeGroupContainer_OffsetX"] = 3,
   ["LargeGroupContainer_OffsetY"] = 140,
+ 
   ["LargeGroup_Width"] = 100,
   ["LargeGroup_Height"] = 36,
   ["LargeGroup_Font"] = "Bebas Neue",
   ["LargeGroup_FontSize"] = 16,
   ["LargeGroup_FontColor"] = {1,1,1,1},
+ 
   ["LargeGroup_Health_BarColor"] = {0.831372549,0.149019608,0.149019608,1},
-  ["LargeGroup_Health_Texture"] = "Round"
+  ["LargeGroup_Health_Texture"] = "Round",
+
+  -- Target
+  ["TargetContainer_OffsetX"] = 810,
+  ["TargetContainer_OffsetY"] = 80,
+ 
+  ["Target_Width"] = 300,
+  ["Target_Font"] = "Bebas Neue",
+  ["Target_FontSize"] = 16,
+  ["Target_FontColor"] = {1,1,1,1},
+  ["Target_FontColor_Friendly"] = {0,1,0,1},
+  ["Target_FontColor_Hostile"] =  {1,0,0,1},
+  ["Target_FontColor_Interact"] = {1,1,0,1},
+ 
+  ["Target_Health_Height"] = 36,
+  ["Target_Health_BarColor"] = {0.831372549,0.149019608,0.149019608,1},
+  ["Target_Health_Texture"] = "Round",
+  ["Target_Health_Font"] = "Bebas Neue",
+  ["Target_Health_FontSize"] = 14, 
+  ["Target_Health_FontColor"] = {1,1,1,1}
 }