+++ ggFrames/unit/Controls.lua
@@ -8,15 +8,16 @@ GGF.classTextures = {
 -- Create Player Frames
 function GGF.Unit:Controls()
   self.template = GGF.Theme.Load(self.unitName)   -- Load Template (For Rendering)
-  
+
   -- Draw Main Control
   self.frames.main = GGF.Window:CreateBackDrop("GGF_"..self.unitName.."Frame", self.parent, self.template)
 
   -- Draw Name and Class (Level)
   self.frames.nameLb  = GGF.Window:CreateLabel("GGF_"..self.unitName.."NameLB",  self.frames.main, self.template.Name)
-  if self.template.Level ~= false then self.frames.levelLb = GGF.Window:CreateLabel("GGF_"..self.unitName.."LevelLB",  self.frames.nameLb, self.template.Level) end
+  if self.template.Level ~= false then self.frames.levelLb = GGF.Window:CreateLabel("GGF_"..self.unitName.."LevelLb",  self.frames.nameLb, self.template.Level) end
   if self.template.Class ~= false then self.frames.classTx = GGF.Window:CreateTexture("GGF_"..self.unitName.."ClassTx", self.frames.main, self.template.Class) end
   if self.template.Leader ~= false then self.frames.leaderTx = GGF.Window:CreateTexture("GGF_"..self.unitName.."LeadTx", self.frames.main, self.template.Leader) end
+  if self.template.Caption ~= false then self.frames.captionLb = GGF.Window:CreateLabel("GGF_"..self.unitName.."CaptionLb", self.frames.main, self.template.Caption) end
 
   -- Misc Labels
   self.frames.death     = GGF.Window:CreateBackDrop("GGF_"..self.unitName.."Death", self.frames.main, self.template.Death)
@@ -63,7 +64,6 @@ function GGF.Unit:Controls()
     if self.template.Experience.Label ~= false then
       self.frames.experienceLb   = GGF.Window:CreateLabel("GGF_"..self.unitName.."ExperienceLabel", self.frames.experienceBd, self.template.Experience.Label)
       self.frames.experienceLb:SetHidden(false)
-      GGF.Debug:New("Created Exp Label", self.frames.experienceLb)
     elseif self.frames.experienceLb ~= nil then
       self.frames.experienceLb:SetHidden(true)
     end
@@ -72,6 +72,9 @@ function GGF.Unit:Controls()
     self.frames.experienceBd:SetHidden(true)
   end
 
+  -- Click Handler
+  -- self.frames.main:SetMouseEnabled(true)
+  -- self.frames.main:SetHandler("OnMouseUp", function(self, btn, upInside) d("Clicking on unit frames will be coming soon") end)
 end
 
 function GGF.Unit:SetPosition()